add delte password pop up

+++ src/app/components/pages/vaults/vaults-page/vault/vault.component.ts
@@ -1,6 +1,7 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { VaultService } from './services/vault.service';
 import { Vault } from '../../../../../models/vault';
+import { ModalService } from '../../../../../services/ui/modal.service';
 
 @Component({
   selector: 'app-vault',
@@ -10,11 +11,30 @@ import { Vault } from '../../../../../models/vault';
 })
 export class VaultComponent {
   vault$ = this.vaultService.vault$;
-  constructor(private vaultService: VaultService) {
+  constructor(
+    private vaultService: VaultService,
+    private modalService: ModalService
+  ) {
     this.vault$ = this.vaultService.vault$;
   }
 
   deleteVault(id: Vault['id']): void {
-    this.vaultService.deleteVault(id);
+    this.modalService
+      .openModal({
+        title: 'Delete vault',
+        body: 'Are you sure you want to delete this vault?',
+        confirmButtonText: 'Delete',
+        confirmButtonClass: 'btn-success',
+        cancelButtonText: 'Cancel',
+        cancelButtonClass: 'btn-danger',
+      })
+      .then((result: boolean): void => {
+        if (result) {
+          this.vaultService.deleteVault(id);
+        }
+      })
+      .catch((error: number): void => {
+        console.log('Error:', error);
+      });
   }
 }
